Show progress as we transcode and copy files
We use ANSI codes to overwrite a previous "in progress" message
with a message stating that the operation is finished. A floating
line, such as a progress line, is displayed below all non-floating
lines, and can be "unfloated", i.e., turned into a regular printed
line that doesn't float below anything.